AuthLogic-to-Datalog IR.
This commit involves just partial progress because the laptop with which
it is written is on its last breath. An IR between the authorization
logic is written. The souffle emitter is updated to take ASTs from the
datalog IR as inputs. The pass that lowers authorization logic into
datalog is not completely done.