Fix antlr build, hide antlr-generated rust code
Now antlr is properly called to re-generate the rust code whenever cargo
build is run. Helpful notes about how to debug the build script and the
antlr code are added to build.rs. The antlr-generated code is better
encapsulated from the human-written code, and the antlr-generated code
is hidden via the gitignore.