Implement queries
Queries are implemented in the surface language to check if a fact (with
no conditions) is true. Souffle does not have queries, but a translation
from queries into datalog is implemented. Instead, souffle relations can
be emitted as CSV files. Essentially a query of the form:

query_name principal says fact_name(<args>) ?

is converted into the souffle program fragment:

query_name("dummy_var") :- says_fact_name(principal, <args>),
    grounded_dummy("dummy_var").
grounded_dummy("dummy_var").

and the check is done by seeing if the CSV for query_name is empty or
not.